requirements.txt
aerich==0.7.2 ; python_version >= "3.11" and python_version < "4.0"
aiosqlite==0.17.0 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.3.0 ; python_version >= "3.11" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.11" and python_version < "3.12.0"
asyncpg==0.29.0 ; python_version >= "3.11" and python_version < "4.0"
bcrypt==4.1.2 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
cryptography==42.0.3 ; python_version >= "3.11" and python_version < "4.0"
dictdiffer==0.9.0 ; python_version >= "3.11" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.109.2 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.6 ; python_version >= "3.11" and python_version < "4.0"
iso8601==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
passlib[bcrypt]==1.7.4 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy"
pydantic-core==2.16.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.6.1 ; python_version >= "3.11" and python_version < "4.0"
pypika-tortoise==0.1.6 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
python-jose[cryptography]==3.3.0 ; python_version >= "3.11" and python_version < "4.0"
python-multipart==0.0.9 ; python_version >= "3.11" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.36.3 ; python_version >= "3.11" and python_version < "4.0"
tomlkit==0.12.3 ; python_version >= "3.11" and python_version < "4.0"
tortoise-orm==0.20.0 ; python_version >= "3.11" and python_version < "4.0"
tortoise-orm[asyncpg]==0.20.0 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.27.1 ; python_version >= "3.11" and python_version < "4.0"
